backend/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.11.9",
    "google-genai>=1.62.0",
    "google-generativeai>=0.8.6",
    "mediapipe>=0.10.32",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
